[#7] refactor : 컴포넌트 경로 변경 및 리뷰사항 반영
