+++ imgoing-frontend/src/screens/MainScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components/native';
-import HomeLanding from '../components/HomeLanding';
+import HomeLanding from '../components/HomeLanding/HomeLanding';
 
 const MainScreen = () => {
   return (
@@ -17,4 +17,5 @@ const Wrapper = styled.SafeAreaView`
   align-items: center;
   justify-content: center;
   background-color: ${({ theme }) => theme.colors.white};
+  width: 100%;
 `;
